requirements.txt
alembic==1.11.1
APScheduler==3.10.1
blinker==1.6.2
cffi==1.15.1
click==8.1.6
colorama==0.4.6
cryptography==41.0.2
decorator==5.1.1
Flask==2.3.2
Flask-APScheduler==1.12.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-Reuploaded==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
packaging==23.1
Pillow==10.0.0
psutil==5.9.5
pycparser==2.21
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.21
style==1.1.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
update==0.0.1
validators==0.20.0
Werkzeug==2.3.6
zipp==3.16.2
flask_wtf==1.2.1
wtforms~=3.1.1
Flask-Session==0.5.0